requirements.txt
astroid==2.7.3
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
html5lib==1.1
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
jupyter-client==7.0.3
jupyter-core==4.8.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lxml==4.6.3
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pandas-datareader==0.10.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prompt-toolkit==3.0.20
pyalgotrading==2021.8.1
Pygments==2.10.0
pylint==2.10.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pyzmq==22.3.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
